lnd: move initNeutrinoBackend to lnd.go
